+++ src/views/header.php
@@ -44,8 +44,14 @@
             <link rel="stylesheet" href="../assets/css/pages/cadastroProjeto.css">
         <?php elseif ($page == 'registraProjeto'): ?>
             <link rel="stylesheet" href="../assets/css/pages/registraProjeto.css">
+        <?php elseif ($page == 'excluiProjeto'): ?>
+            <link rel="stylesheet" href="../assets/css/pages/excluiProjeto.css">
         <?php elseif ($page == 'cadastroAluno'): ?>
             <link rel="stylesheet" href="../assets/css/pages/cadastroAluno.css">
+        <?php elseif ($page == 'revisaComentarios'): ?>
+            <link rel="stylesheet" href="../assets/css/pages/revisaComentarios.css">
+        <?php elseif ($page == 'editarProjeto'): ?>
+            <link rel="stylesheet" href="../assets/css/pages/editarProjeto.css">
         <?php endif; ?>
     <?php endif; ?>
 </head>